build: don't use lint-ci on Travis
The `lint-ci` Makefile target differs from `lint` in that it writes to
a tap file and not stdout and also stops execution when an error is
found (e.g. if JavaScript linting fails the C++ and docs linting are
not run).

The switch to `lint-ci` was to enable Python linting. Revert to `lint`
and add the `lint-py` target.

PR-URL: https://github.com/nodejs/node/pull/27062
Reviewed-By: Refael Ackermann <refack@gmail.com>
Reviewed-By: Ruben Bridgewater <ruben@bridgewater.de>